AudioPlayer: Removed unused Start and Stop methods
